fix: support large models and partial biases
- Change n_parameters from int to size_t to prevent overflow for 3B+ models
- Add partial bias support in convert.py for models like Qwen2.5 (q/k/v bias but no o_proj bias)
- Fix tokenizer fallback in run.py for models with slow tokenizers
- Fix temp directory cleanup in run.py to check output before cleanup

+++ src/baremetal/model.h
@@ -1,12 +1,13 @@
 #ifndef BMT_MODEL_H
 #define BMT_MODEL_H
 
+#include <stddef.h>
 #include "baremetal.h"
 
 struct bm_model_s {
     bm_arch_t arch;
 
-    int      n_parameters;
+    size_t   n_parameters;
 
     int      kv_dim;
     int      head_size;
